pyproject.toml
[project]
name = "ltai-news"
version = "0.0.1"
description = "Daily AI-powered digest of the latest videos from YouTube channels for the Let's Talk AI ecosystem"
authors = [
    {name = "Thomas Bustos"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.0",
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.3",
    "langchain-google-genai>=3.2.0",
    "loguru>=0.7.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-requests>=0.45b0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.25.0",
    "uvicorn[standard]>=0.38.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
]
